Support of Symfony bundles in paths
Some assets could not be in a place indicated in baseUrl. This should be taken into consideration while receiving module name from asset. If an original file is not in a place indicated in baseUrl, it means the path to the file can be found in 'paths'. That's where the name of the module can be taken of.